typo, namespaces, remove flushing intermediate data, rename
